+++ app/Http/Controllers/User/ProductController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\User\Product;
+namespace App\Http\Controllers\User;
 
 use App\Enums\ProductStatus;
 use App\Http\Controllers\Controller;
@@ -8,6 +8,7 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Spatie\Tags\Tag;
+use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
 {
@@ -23,21 +24,102 @@ class ProductController extends Controller
         return view('user.products.check', compact('userProducts', 'message'));
     }
 
-    public function demoteData(Product $product)
+    public function create()
     {
-        // 根據當前狀態切換到相反的狀態
-        $newStatus = $product->status === ProductStatus::Active
-            ? ProductStatus::Inactive
-            : ProductStatus::Active;
+        $tags = Tag::whereNull('deleted_at')->get();
+        return view('user.products.create', compact('tags'));
+    }
 
-        // 更新商品的狀態
-        $product->update([
-            'status' => $newStatus,
-        ]);
+    public function store(Request $request)
+    {
+        // 基本驗證規則
+        $rules = [
+            'name' => ['required', 'string', 'max:50'],
+            'price' => ['required', 'numeric', 'min:0','max:9999'],
+            'description' => ['required', 'string'],
+            'grade' => ['required', 'string', 'not_in:選擇適用的年級...'],
+            'semester' => ['required', 'string', 'not_in:選擇學期...'],
+            'category' => ['required', 'string', 'not_in:選擇課程類別...'],
+            'images' => ['required', 'array', 'min:1', 'max:5'],
+            'images.*' => [
+                'required',
+                'image',
+                'mimes:svg,png,jpg,jpeg,gif',
+                'max:2048',
+                'dimensions:max_width=3200,max_height=3200'
+            ],
+        ];
 
-        $message = "商品{$newStatus->label()}！";
+        try {
+            // 驗證
+            $validated = $request->validate($rules);
 
-        return redirect()->route('user.products.index')->with('success', $message);
+            $product = Product::create([
+                'name' => $validated['name'],
+                'price' => $validated['price'],
+                'description' => $validated['description'],
+                'user_id' => auth()->id(),
+            ]);
+
+            // 處理圖片上傳
+            if ($request->hasFile('images')) {
+                foreach ($request->file('images') as $index => $image) {
+                    if ($index >= 5) break;
+                    $product->addMedia($image)->toMediaCollection('images');
+                }
+            }
+
+            // 處理標籤
+            $tagTypes = [
+                ['type' => '年級', 'slug' => $request->input('grade')],
+                ['type' => '學期', 'slug' => $request->input('semester')],
+                ['type' => '課程', 'slug' => $request->input('category')]
+            ];
+
+            foreach ($tagTypes as $tagType) {
+                $tag = Tag::where('slug->zh', $tagType['slug'])
+                         ->where('type', $tagType['type'])
+                         ->first();
+
+                if ($tag) {
+                    $product->attachTag($tag);
+                }
+            }
+
+            return redirect()->route('user.products.create')->with('success', '產品已成功創建！');
+        } catch (\Illuminate\Validation\ValidationException $e) {
+            return redirect()->back()
+                ->withErrors($e->validator)
+                ->withInput();
+        }
+    }
+
+    public function edit(Request $request, Product $product)
+    {
+        $gradeTag = $product->tags->firstWhere('type', '年級');
+        $semesterTag = $product->tags->firstWhere('type', '學期');
+        $categoryTag = $product->tags->firstWhere('type', '課程');
+        $tags = Tag::whereNull('deleted_at')->get();
+
+        if ($request->hasFile('images')) {
+            $images = $request->file('images');
+
+            // 先獲取並按 id 升序排序已存在的圖片
+            $existingMedia = $product->getMedia('images')->sortBy('id')->values();
+
+            foreach ($images as $index => $image) {
+                // 確認是否需要替換該圖片
+                if (isset($existingMedia[$index])) {
+                    // 如果圖片已存在，則替換
+                    $existingMedia[$index]->delete();
+                }
+
+                // 上傳新的圖片
+                $product->addMedia($image)->toMediaCollection('images');
+            }
+        }
+
+        return view('user.products.edit', compact('product', 'tags', 'gradeTag', 'semesterTag', 'categoryTag'));
     }
 
     public function update(Request $request, Product $product)
@@ -164,5 +246,48 @@ class ProductController extends Controller
         // 重新導向到產品清單頁面，並標註成功訊息
         return redirect()->route('user.products.index')->with('success', '產品已成功刪除');
     }
+
+    public function uploadTempImage(Request $request)
+    {
+        $request->validate([
+            'image' => 'required|image|max:2048',
+            'position' => 'required|integer|min:0|max:4'
+        ]);
+
+        try {
+            $file = $request->file('image');
+            $path = $file->store('temp/products', 'public');
+
+            return response()->json([
+                'success' => true,
+                'path' => $path,
+                'position' => $request->position
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => '上傳圖片時發生錯誤'
+            ]);
+        }
+    }
+
+    public function demoteData(Product $product)
+    {
+        // 根據當前狀態切換到相反的狀態
+        $newStatus = $product->status === ProductStatus::Active
+            ? ProductStatus::Inactive
+            : ProductStatus::Active;
+
+        // 更新商品的狀態
+        $product->update([
+            'status' => $newStatus,
+        ]);
+
+        $message = "商品{$newStatus->label()}！";
+
+        return redirect()->route('user.products.index')->with('success', $message);
+    }
+
+
 
 }
