+++ routes/user.php
@@ -1,8 +1,6 @@
 <?php
 
-use App\Http\Controllers\User\Product\CreateController;
-use App\Http\Controllers\User\Product\EditController;
-use App\Http\Controllers\User\Product\ProductController;
+use App\Http\Controllers\User\ProductController;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('user')->name('user.')->middleware(['auth'])->group(function () {
@@ -13,7 +11,7 @@ Route::prefix('user')->name('user.')->middleware(['auth'])->group(function () {
     Route::put('/products/{product}/demote', [ProductController::class, 'demoteData'])
         ->name('products.demoteData');
     // 商品修改頁面
-    Route::get('/products/{product}/edit', [EditController::class, 'edit'])
+    Route::get('/products/{product}/edit', [ProductController::class, 'edit'])
         ->name('products.edit');
     // 商品>修改 資料
     Route::put('/products/{product}', [ProductController::class, 'update'])
@@ -24,10 +22,10 @@ Route::prefix('user')->name('user.')->middleware(['auth'])->group(function () {
         ->name('products.destroy');
 
     // 刊登商品頁面
-    Route::get('/products/create', [CreateController::class, 'create'])
+    Route::get('/products/create', [ProductController::class, 'create'])
         ->name('products.create');
     // 商品建立資料
-    Route::post('/products', [CreateController::class, 'store'])
+    Route::post('/products', [ProductController::class, 'store'])
         ->name('products.store');
 
 });
